WIP invoker: Add -S/--systemd to notify systemd of the boosted PID
XXX If the booster isn't double-forked, systemd seems to have trouble
following the boosted process' state.

XXX Think a bit about systemd Type. notify and dbus should be available
for the child process to indicate, forking is probably the right
behavior for most cases, and simple is technically wrong. I think.

XXX Setting MAINPID does not change the cgroup of the boosted process.